requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.0.3
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-google-genai==2.1.3
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.31
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
pydeck==0.9.1
pygments==2.19.1
pymupdf==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.40
stack-data==0.6.3
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
yarl==1.19.0
zstandard==0.23.0
